Rudimentary polynomial root solver.

+++ arduino-pico/ToyStacky/TS-core-fn1Param.h
@@ -100,7 +100,7 @@ bool fn1ParamScalar(Machine* vm, const char* fnname, int fnindex, int isTrig, in
 		//convert back to polar mode
 		if (!alm0double(c.real)) {
 			long double temp = abso(c);
-			c.imag = atan(c.imag/c.real);
+			c.imag = atanl(c.imag/c.real);
 			c.real = temp;
 		} else {
 			c.imag = 1.570796326794896619231L;//pi/2
